refactor: use Angular Material new MDC implementations
Using the new implementation required a new `mdc-overrides.css` stylesheet to fix styling issues with Tailwind.

The custom theme we previously created but never used was also replaced with one of the prebuilt themes included with the library.

No other major changes were needed when replacing the legacy components with the new ones.
