+++ angular.json
@@ -23,7 +23,11 @@
             "polyfills": "src/polyfills.ts",
             "tsConfig": "tsconfig.app.json",
             "assets": ["src/favicon.ico", "src/assets"],
-            "styles": ["src/custom-theme.scss", "src/styles.css"],
+            "styles": [
+              "node_modules/@angular/material/prebuilt-themes/indigo-pink.css",
+              "src/styles.css",
+              "src/mdc-overrides.css"
+            ],
             "scripts": []
           },
           "configurations": {
@@ -99,7 +103,8 @@
     }
   },
   "cli": {
-    "schematicCollections": ["@angular-eslint/schematics"]
+    "schematicCollections": ["@angular-eslint/schematics"],
+    "analytics": "14912cdc-d6cf-423b-847b-7202aa0d2139"
   },
   "schematics": {
     "@angular-eslint/schematics:application": {
