+++ src/styles.css
@@ -5,7 +5,7 @@
 *,
 *::before,
 *::after {
-  @apply m-0 relative font-normal box-border;
+  @apply m-0 font-normal box-border;
 }
 
 body {
